CRUD usuario

+++ src/Controller/EventoController.java
@@ -12,11 +12,11 @@ public class EventoController extends Controller {
 
     public static double getPorcentagemCred(int eventoId) {
         int presencas = CredenciamentoController.presencasConfirmadas(eventoId), inscricoes = InscricaoController.inscricoesRealizadas(eventoId);
-        return inscricoes==0? 0: presencas/inscricoes;
+        return inscricoes==0? 0: (double)presencas/(double)inscricoes;
     }
     
     public static double getPorcentagemCred() {
         int presencas = CredenciamentoController.presencasConfirmadas(), inscricoes = InscricaoController.inscricoesRealizadas();
-        return inscricoes==0? 0: presencas/inscricoes;
+        return inscricoes==0? 0: (double)presencas/(double)inscricoes;
     }
 }
